+++ server/service/src/programs/mod.rs
@@ -1,8 +1,6 @@
 pub mod contact_trace;
 pub mod encounter;
-pub mod indicator_value;
 pub mod patient;
 pub mod program_enrolment;
 pub mod program_event;
-pub mod program_indicator;
 pub mod update_program_document;
