+++ server/service/src/service_provider.rs
@@ -37,15 +37,17 @@ use crate::{
     programs::{
         contact_trace::{ContactTraceService, ContactTraceServiceTrait},
         encounter::{EncounterService, EncounterServiceTrait},
-        indicator_value::{IndicatorValueService, IndicatorValueServiceTrait},
         patient::{PatientService, PatientServiceTrait},
         program_enrolment::{ProgramEnrolmentService, ProgramEnrolmentServiceTrait},
         program_event::{ProgramEventService, ProgramEventServiceTrait},
-        program_indicator::{ProgramIndicatorService, ProgramIndicatorServiceTrait},
     },
     repack::{RepackService, RepackServiceTrait},
     report::report_service::{ReportService, ReportServiceTrait},
-    requisition::{RequisitionService, RequisitionServiceTrait},
+    requisition::{
+        indicator_value::{IndicatorValueService, IndicatorValueServiceTrait},
+        program_indicator::{ProgramIndicatorService, ProgramIndicatorServiceTrait},
+        RequisitionService, RequisitionServiceTrait,
+    },
     requisition_line::{RequisitionLineService, RequisitionLineServiceTrait},
     rnr_form::{RnRFormService, RnRFormServiceTrait},
     sensor::{SensorService, SensorServiceTrait},
